# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Challenge Arena API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{userId}/update:
    put:
      operationId: updateUser
      tags: [users]
      summary: Update user profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        freefireUid:
          type: string
          nullable: true
        ign:
          type: string
          nullable: true
        gender:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        email:
          type: string
          nullable: true
        whatsappPhone:
          type: string
          nullable: true
        isAdmin:
          type: boolean
          description: True when user may access admin API (env or Firestore flag)
        stats:
          $ref: "#/components/schemas/PlayerStats"
          nullable: true
      required:
        - id
        - username
        - createdAt
    UpdateUserBody:
      type: object
      properties:
        freefireUid:
          type: string
        ign:
          type: string
        gender:
          type: string
          enum: [male, female, other]
          nullable: true
...
    PlayerStats:
      type: object
      properties:
        userId:
          type: string
        matchesPlayed:
          type: integer
        wins:
          type: integer
        losses:
          type: integer
        winStreak:
          type: integer
        weeklyWins:
          type: integer
      required:
        - userId
        - matchesPlayed
        - wins
        - losses
        - winStreak
        - weeklyWins
...